well now it kinda works with 500k starting tree, i have never seen struct greater in practice, so i will stop. yes it is still not optimized, but it was fun. and also it is fun in how and why now it works

+++ structuresOfAllKind/structuresOfAllKind/MyBinaryTree.h
@@ -11,7 +11,7 @@ public:
 	void add(long long value);
 	BinaryTreeNode* find(long long key);
 	//родилась из неправильного понимания теории, решил сделать потому что забавно
-	bool insert(int key, int value);
+	bool insert(long long key, long long value);
 	void remove(int key);
 	void traverse(void* foo);
 	void join(MyBinaryTree anotherTree);
